Minor fixes

+++ src/main/java/com/qa/birdNoise/repo/BirdNoiseRepo.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.qa.birdNoise.domain.BirdNoise;
 
-public interface BirdNoiseRepo extends JpaRepository<BirdNoise, Long> {
-	
+public interface BirdNoiseRepo extends JpaRepository<BirdNoise, Integer> {
+
 	List<BirdNoise> findByPoster(String poster);
-	
+
 }
