+++ src/main/java/com/qa/birdNoise/service/ServiceInterface.java
@@ -3,11 +3,11 @@ package com.qa.birdNoise.service;
 import java.util.List;
 
 public interface ServiceInterface<T> {
-	
+
 	public T create (T t);
 	public List<T> getAll();
-	public T get(Long id);
+	public T get(Integer id);
 	public boolean replace(T t);
-	public boolean remove(Long id);
+	public boolean remove(Integer id);
 
 }
